+++ client/context.go
@@ -208,10 +208,10 @@ func (ctx Context) PrintString(str string) error {
 	return err
 }
 
-// PrintObject outputs toPrint to the ctx.Output based on ctx.OutputFormat which is
+// PrintProto outputs toPrint to the ctx.Output based on ctx.OutputFormat which is
 // either text or json. If text, toPrint will be YAML encoded. Otherwise, toPrint
 // will be JSON encoded using ctx.JSONMarshaler. An error is returned upon failure.
-func (ctx Context) PrintObject(toPrint proto.Message) error {
+func (ctx Context) PrintProto(toPrint proto.Message) error {
 	// always serialize JSON initially because proto json can't be directly YAML encoded
 	out, err := ctx.JSONMarshaler.MarshalJSON(toPrint)
 	if err != nil {
@@ -220,8 +220,9 @@ func (ctx Context) PrintObject(toPrint proto.Message) error {
 	return ctx.printOutput(out)
 }
 
-// PrintObjectLegacy is a variant of PrintObject that doesn't require a proto type
-// and uses amino JSON encoding. It will be removed in the near future!
+// PrintObjectLegacy is a variant of PrintProto that doesn't require a proto.Message type
+// and uses amino JSON encoding.
+// Deprecated: It will be removed in the near future!
 func (ctx Context) PrintObjectLegacy(toPrint interface{}) error {
 	out, err := ctx.LegacyAmino.MarshalJSON(toPrint)
 	if err != nil {